refactor: auto fill from file now works as it should
