+++ src/Components/Autofill/autoFillFromFile.js
@@ -1,0 +1,27 @@
+export const autoFillFromFileDynamic = async ({
+  dispatch,
+  index,
+  blockIndex,
+  filteredBlocks
+}) => {
+  if (!filteredBlocks[blockIndex]) return
+
+  let block = filteredBlocks[blockIndex].content
+
+  dispatch({ type: 'autoFill', payload: block, index })
+}
+
+export const autoFillFromFileStatic = async ({
+  dispatch,
+  savedJSON,
+  blockName
+}) => {
+  console.log(savedJSON, blockName)
+  console.log(savedJSON[blockName])
+
+  dispatch({
+    type: 'autoFillBlock',
+    payload: savedJSON[blockName],
+    blockName
+  })
+}
