+++ src/Components/Autofill/index.js
@@ -1,14 +1,6 @@
-import axios from 'axios'
-import React from 'react'
 import styled from 'styled-components'
 
-import { useAppState, useAppDispatch } from '../../context/state-context'
-import { placeholderImage } from '../../lib/placeholderImage'
-import { convertEmojis } from '../../lib/emojiConvert'
-
 import fillIcon from '../../icons/autoFill.svg'
-import { ReactComponent as AutoFillIcon } from '../../icons/autoFill.svg'
-
 const fs = window.require('fs-extra')
 const { dialog } = window.require('electron').remote
 
@@ -56,378 +48,176 @@ export const removeTransformations = url => {
   }
 }
 
-export const autoFillFromFile = async (dispatch, index, value, type) => {
+export const isFromFileCheck = e =>
+  e.target[e.target.selectedIndex].getAttribute('data-file')
+
+export const getFileJSON = async () => {
   const { filePaths } = await dialog.showOpenDialog({
     properties: ['openFile'],
     filters: [{ extensions: ['json'] }]
   })
 
   if (filePaths) {
-    let file = JSON.parse(fs.readFileSync(filePaths[0], 'utf8'))
-
-    console.log(file)
-
-    if (type === 'main' || type === 'lower') {
-      let contentBlocks = file.contentBlocks.filter(block => {
-        return block.type === type
-      })
-
-      let blockIndex =
-        type === 'main'
-          ? parseInt(value.substring(3, 4)) - 1
-          : value.substring(0, 1)
-
-      if (!contentBlocks[blockIndex]) return
-
-      let block = contentBlocks[blockIndex].content
-
-      let imageArr =
-        type === 'main' ? [block.image, block.mobile] : [block.image]
-
-      imageArr.forEach((x, i) => {
-        placeholderImage(x).then(placeholder => {
-          dispatch({
-            type: 'placeholderImage',
-            name: i === 0 ? 'image' : 'mobile',
-            index,
-            payload: placeholder
-          })
-        })
-      })
-      console.log(block, 'this is the block')
-      dispatch({ type: 'autoFill', payload: block, index })
-    } else if (type === 'sale') {
-      let saleContentBlocks = file.contentBlocks.filter(
-        block => block.type === type
-      )
-
-      let saleBlockIndex = value.substring(0, 1)
-
-      if (!saleContentBlocks[saleBlockIndex]) return
-
-      dispatch({
-        type: 'autoFill',
-        payload: saleContentBlocks[saleBlockIndex].content,
-        index
-      })
-    } else if (type.toLowerCase().includes('categories')) {
-      dispatch({
-        type: 'autoFillBlock',
-        payload: file[type],
-        blockName: type
-      })
-    }
+    return JSON.parse(fs.readFileSync(filePaths[0], 'utf8'))
+  } else {
+    return null
   }
 }
 
-let createBlock = (el, type) => {
-  if (type === 'lower') {
-    let subtitle = el.querySelector('.subtitle3')
-    let title = el.querySelector('.title3')
+// export const autoFillFromFile = async (dispatch, index, value, type) => {
+//   const { filePaths } = await dialog.showOpenDialog({
+//     properties: ['openFile'],
+//     filters: [{ extensions: ['json'] }]
+//   })
 
-    return {
-      cta: el.querySelector('button').textContent,
-      image: removeTransformations(el.querySelector('img').dataset.src),
-      subtitle: subtitle ? subtitle.textContent : '',
-      title: title ? title.textContent : '',
-      url: el.getAttribute('href')
-    }
-  } else if (type === 'main') {
-    let srcSet = el.querySelector('source').dataset.srcset
-    let urls = el.querySelectorAll('a')
-    let buttons = el.querySelectorAll('button')
-    let subtitle = el.querySelector('.subtitle1')
-    let title = el.querySelector('.title1')
+//   if (filePaths) {
+//     let file = JSON.parse(fs.readFileSync(filePaths[0], 'utf8'))
 
-    return {
-      image: removeTransformations(el.querySelector('img').dataset.src),
-      mobile: removeTransformations(srcSet.match(/^https:[^ ]+/gi)[0]),
-      primaryCta: buttons[0].textContent,
-      primaryUrl: urls[0].getAttribute('href'),
-      secondaryCta: buttons[1] ? buttons[1].textContent : '',
-      secondaryUrl: urls[1] ? urls[1].getAttribute('href') : '',
-      subtitle: subtitle ? subtitle.textContent : '',
-      svg: el.querySelector('svg').outerHTML,
-      title: title ? title.textContent : ''
-    }
-  } else if (type === 'sale') {
-    let ctas = Array.from(el.querySelectorAll('a')).map(cta => {
-      let url = cta.getAttribute('href')
-      let text = cta.querySelector('button').textContent
+//     console.log(file)
 
-      return {
-        url,
-        text
-      }
-    })
+//     switch (type) {
+//     }
 
-    return {
-      title: el.querySelector('h2').textContent.trim(),
-      ctas
-    }
-  }
-}
+//     if (type === 'main' || type === 'lower') {
+//       let contentBlocks = file.contentBlocks.filter(block => {
+//         return block.type === type
+//       })
 
-let autoFillFromSite = async (
-  dispatch,
-  index,
-  value,
-  territory,
-  type,
-  selector
-) => {
-  if (value === 'default') return
+//       let blockIndex =
+//         type === 'main'
+//           ? parseInt(value.substring(3, 4)) - 1
+//           : value.substring(0, 1)
 
-  let { data } = await axios.get(territory.url)
-  let parser = new DOMParser()
-  let html = parser.parseFromString(data, 'text/html')
+//       if (!contentBlocks[blockIndex]) return
 
-  let el =
-    type === 'main'
-      ? html.querySelector(`.${value}`)
-      : html.querySelectorAll(selector)[value]
+//       let block = contentBlocks[blockIndex].content
 
-  if (el === null) return
+//       let imageArr =
+//         type === 'main' ? [block.image, block.mobile] : [block.image]
 
-  let block = createBlock(el, type)
+//       imageArr.forEach((x, i) => {
+//         placeholderImage(x).then(placeholder => {
+//           dispatch({
+//             type: 'placeholderImage',
+//             name: i === 0 ? 'image' : 'mobile',
+//             index,
+//             payload: placeholder
+//           })
+//         })
+//       })
+//       console.log(block, 'this is the block')
+//       dispatch({ type: 'autoFill', payload: block, index })
+//     } else if (type === 'sale') {
+//       let saleContentBlocks = file.contentBlocks.filter(
+//         block => block.type === type
+//       )
 
-  if (type === 'main' || type === 'lower') {
-    trimWhiteSpace(block)
+//       let saleBlockIndex = value.substring(0, 1)
 
-    let imageArr = type === 'main' ? [block.image, block.mobile] : [block.image]
+//       if (!saleContentBlocks[saleBlockIndex]) return
 
-    imageArr.forEach((x, i) => {
-      placeholderImage(x).then(placeholder => {
-        dispatch({
-          type: 'placeholderImage',
-          name: i === 0 ? 'image' : 'mobile',
-          index,
-          payload: placeholder
-        })
-      })
-    })
-  }
+//       dispatch({
+//         type: 'autoFill',
+//         payload: saleContentBlocks[saleBlockIndex].content,
+//         index
+//       })
+//     } else if (type.toLowerCase().includes('categories')) {
+//       dispatch({
+//         type: 'autoFillBlock',
+//         payload: file[type],
+//         blockName: type
+//       })
+//     }
+//   }
+// }
 
-  dispatch({ type: 'autoFill', payload: block, index })
-}
+// let createBlock = (el, type) => {
+//   if (type === 'lower') {
+//     let subtitle = el.querySelector('.subtitle3')
+//     let title = el.querySelector('.title3')
 
-export const AutoFillContent = ({ index, type }) => {
-  const { territory } = useAppState()
-  let dispatch = useAppDispatch()
+//     return {
+//       cta: el.querySelector('button').textContent,
+//       image: removeTransformations(el.querySelector('img').dataset.src),
+//       subtitle: subtitle ? subtitle.textContent : '',
+//       title: title ? title.textContent : '',
+//       url: el.getAttribute('href')
+//     }
+//   } else if (type === 'main') {
+//     let srcSet = el.querySelector('source').dataset.srcset
+//     let urls = el.querySelectorAll('a')
+//     let buttons = el.querySelectorAll('button')
+//     let subtitle = el.querySelector('.subtitle1')
+//     let title = el.querySelector('.title1')
 
-  let mainPopulate = async (territory, index, e) => {
-    if (e.target.value === 'default') return
+//     return {
+//       image: removeTransformations(el.querySelector('img').dataset.src),
+//       mobile: removeTransformations(srcSet.match(/^https:[^ ]+/gi)[0]),
+//       primaryCta: buttons[0].textContent,
+//       primaryUrl: urls[0].getAttribute('href'),
+//       secondaryCta: buttons[1] ? buttons[1].textContent : '',
+//       secondaryUrl: urls[1] ? urls[1].getAttribute('href') : '',
+//       subtitle: subtitle ? subtitle.textContent : '',
+//       svg: el.querySelector('svg').outerHTML,
+//       title: title ? title.textContent : ''
+//     }
+//   } else if (type === 'sale') {
+//     let ctas = Array.from(el.querySelectorAll('a')).map(cta => {
+//       let url = cta.getAttribute('href')
+//       let text = cta.querySelector('button').textContent
 
-    if (e.target.value.includes('file')) {
-      autoFillFromFile(dispatch, index, e.target.value, 'main')
-    } else {
-      autoFillFromSite(dispatch, index, e.target.value, territory, 'main')
-    }
-  }
+//       return {
+//         url,
+//         text
+//       }
+//     })
 
-  let lowerPopulate = async (territory, index, e) => {
-    e.persist()
+//     return {
+//       title: el.querySelector('h2').textContent.trim(),
+//       ctas
+//     }
+//   }
+// }
 
-    if (e.target.value.includes('file')) {
-      autoFillFromFile(dispatch, index, e.target.value, 'lower')
-    } else {
-      autoFillFromSite(
-        dispatch,
-        index,
-        e.target.value,
-        territory,
-        'lower',
-        '.slick-three > div > a'
-      )
-    }
-  }
+// let autoFillFromSite = async (
+//   dispatch,
+//   index,
+//   value,
+//   territory,
+//   type,
+//   selector
+// ) => {
+//   if (value === 'default') return
 
-  let salePopulate = async (territory, index, e) => {
-    if (e.target.value === 'default') return
+//   let { data } = await axios.get(territory.url)
+//   let parser = new DOMParser()
+//   let html = parser.parseFromString(data, 'text/html')
 
-    if (e.target.value.includes('file')) {
-      autoFillFromFile(dispatch, index, e.target.value, 'sale')
-    } else {
-      autoFillFromSite(
-        dispatch,
-        index,
-        e.target.value,
-        territory,
-        'sale',
-        '.categories-sale__container'
-      )
-    }
-  }
+//   let el =
+//     type === 'main'
+//       ? html.querySelector(`.${value}`)
+//       : html.querySelectorAll(selector)[value]
 
-  let selectSwitch = (e, type) => {
-    switch (type) {
-      case 'main':
-        return mainPopulate(territory, index, e)
-      case 'lower':
-        return lowerPopulate(territory, index, e)
-      case 'sale':
-        return salePopulate(territory, index, e)
-      default:
-        console.log('type not handled')
-    }
-  }
+//   if (el === null) return
 
-  return (
-    <IconContent>
-      <select
-        name="autofill"
-        value="default"
-        onChange={e => selectSwitch(e, type)}
-      >
-        {type === 'main' && (
-          <>
-            <option value="default">No Row</option>
-            <option value="row1">Row 1</option>
-            <option value="row2">Row 2</option>
-            <option value="row3">Row 3</option>
-            <option value="row4">Row 4</option>
-            <option value="row1 file">Row 1 (from file)</option>
-            <option value="row2 file">Row 2 (from file)</option>
-            <option value="row3 file">Row 3 (from file)</option>
-            <option value="row4 file">Row 4 (from file)</option>
-          </>
-        )}
-        {type === 'lower' && (
-          <>
-            <option value="default">No Lower</option>
-            <option value="0">Lower 1</option>
-            <option value="1">Lower 2</option>
-            <option value="2">Lower 3</option>
-            <option value="3">Lower 4</option>
-            <option value="0 file">Lower 1 (from file)</option>
-            <option value="1 file">Lower 2 (from file)</option>
-            <option value="2 file">Lower 3 (from file)</option>
-            <option value="3 file">Lower 4 (from file)</option>
-          </>
-        )}
-        {type === 'sale' && (
-          <>
-            <option value="default">No Lower</option>
-            <option value="0">CTA's 1</option>
-            <option value="1">CTA's 2</option>
-            <option value="0 file">CTA's 1 (from file)</option>
-            <option value="1 file">CTA's 2 (from file)</option>
-          </>
-        )}
-      </select>
-    </IconContent>
-  )
-}
+//   let block = createBlock(el, type)
 
-const Icon = styled(AutoFillIcon)`
-  width: 28px;
-  height: 28px;
-  cursor: pointer;
-`
+//   if (type === 'main' || type === 'lower') {
+//     trimWhiteSpace(block)
 
-export const AutoFillCategories = () => {
-  const dispatch = useAppDispatch()
-  const { territory } = useAppState()
+//     let imageArr = type === 'main' ? [block.image, block.mobile] : [block.image]
 
-  const handleClick = async (territory, e) => {
-    if (e.target.value.includes('file')) {
-      autoFillFromFile(dispatch, null, e.target.value, 'categories')
-    } else {
-      let { data } = await axios.get(territory.url)
-      let parser = new DOMParser()
-      let html = parser.parseFromString(data, 'text/html')
+//     imageArr.forEach((x, i) => {
+//       placeholderImage(x).then(placeholder => {
+//         dispatch({
+//           type: 'placeholderImage',
+//           name: i === 0 ? 'image' : 'mobile',
+//           index,
+//           payload: placeholder
+//         })
+//       })
+//     })
+//   }
 
-      let categoryTiles = Array.from(
-        html.querySelectorAll('.category-tile__link')
-      )
-
-      if (!categoryTiles.length) return
-
-      let categoriesArr = []
-
-      categoryTiles.forEach(category => {
-        let image = removeTransformations(
-          category.querySelector('.category-tile__image').dataset.src
-        )
-        let url = category.getAttribute('href')
-        let title = category.querySelector('.category-tile__heading')
-          .textContent
-
-        categoriesArr.push({
-          image,
-          url,
-          title
-        })
-
-        categoriesArr.forEach(x => {
-          trimWhiteSpace(x)
-        })
-
-        dispatch({
-          type: 'autoFillBlock',
-          payload: categoriesArr,
-          blockName: 'categories'
-        })
-      })
-    }
-  }
-
-  return (
-    <IconContent>
-      <select
-        name="autofill"
-        value="default"
-        className="categories"
-        onChange={e => handleClick(territory, e)}
-      >
-        <>
-          <option value="default">No categories</option>
-          <option value="0">From site</option>
-          <option value="0 file">From file</option>
-        </>
-      </select>
-    </IconContent>
-  )
-}
-
-export const AutoFillPromos = props => {
-  const dispatch = useAppDispatch()
-  const { territory } = useAppState()
-  const handleClick = async e => {
-    e.persist()
-
-    let { data } = await axios.get(territory.url)
-    let parser = new DOMParser()
-    let html = parser.parseFromString(data, 'text/html')
-
-    let promos = Array.from(html.querySelectorAll('.info'))
-
-    if (!promos.length) return
-
-    let promosArr = []
-
-    promos.forEach(promo => {
-      let url = promo.querySelector('a').getAttribute('href')
-      let title = promo.querySelector('h3').innerHTML
-
-      title = convertEmojis(title)
-
-      promosArr.push({
-        url,
-        title
-      })
-
-      promosArr.forEach(x => {
-        trimWhiteSpace(x)
-      })
-
-      dispatch({
-        type: 'autoFillBlock',
-        payload: promosArr,
-        blockName: 'promoBlocks'
-      })
-    })
-  }
-
-  return <Icon onClick={handleClick} />
-}
+//   dispatch({ type: 'autoFill', payload: block, index })
+// }
