+++ src/Components/DynamicBlocks/Block.js
@@ -3,7 +3,7 @@ import { useAppDispatch } from '../../context/state-context'
 
 import { ReactComponent as DragIcon } from '../../icons/dragIcon.svg'
 
-import { AutoFillContent } from '../Autofill'
+import AutoFillDynamicBlocks from '../Autofill/AutoFillDynamicBlocks'
 import HeroBlock from '../Hero/HeroBlock'
 import LowerBlock from '../Lower/LowerBlock'
 import SaleContainer from '../Sale/SaleContainer'
@@ -79,7 +79,7 @@ const DynamicBlock = ({ block, index }) => {
         >
           Remove
         </Remove>
-        <AutoFillContent index={index} type={block.type} />
+        <AutoFillDynamicBlocks index={index} type={block.type} />
         <DragIcon />
       </DropDown>
       {block.type === 'main' && <HeroBlock index={index} block={block} />}
