+++ src/Components/SaleCategories/SaleCategories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useAppState } from '../../context/state-context'
-import AutofillSaleCategories from '../Autofill/AutofillSaleCategories/index'
+import AutofillSaleCategories from '../Autofill/AutofillSaleCategories'
 
 import DragDrop from '../DragDrop'
 import SaleCategoryBlock from './SaleCategoryBlock'
